api/src/factories/embedding_model_factories/embedding_model_abstract_factory.py
from abc import ABC, abstractmethod

from langchain_core.embeddings import Embeddings


class EmbeddingModelAbstractFactory(ABC):
    @abstractmethod
    def create(self) -> Embeddings:
        raise NotImplementedError
